pyproject.toml
[project]
name = "law-assistant-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "openai>=2.30.0",
    "pinecone>=3.0.0,<4.0.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.43.0",
]
